Add cache validity checking to HDDInfoCacheDataset
This is useful if there are going to be code or data changes

+++ vidlu/data/dataset.py
@@ -19,6 +19,7 @@ from torchvision.datasets.utils import download_and_extract_archive
 import dataclasses as dc
 from functools import cached_property
 from enum import Enum
+from functools import partial
 
 import numpy as np
 from torch.utils.data.dataset import ConcatDataset
@@ -621,32 +622,44 @@ class InfoCacheDataset(Dataset):  # lazy
 
 
 class HDDInfoCacheDataset(InfoCacheDataset):  # TODO
-    def __init__(self, dataset, name_to_func, cache_dir, recompute=False, **kwargs):
+    def __init__(self, dataset, name_to_func, cache_dir, recompute=False, simplify_dataset=None,
+                 **kwargs):
         super().__init__(dataset, name_to_func, **kwargs)
         self.cache_dir = Path(cache_dir)
         self.cache_file = to_valid_path(self.cache_dir / "info_cache" / self.identifier)
         self.cache_file.parent.mkdir(parents=True, exist_ok=True)
         self.recompute = recompute
+        self.check_data = None
+        if simplify_dataset is not None:
+            self.check_data = InfoCacheDataset(simplify_dataset(dataset), name_to_func)
+
+    def _compute_check_data(self):
+        return None if self.check_data is None else self.check_data._compute()
 
     def _compute(self):
-        if self.recompute and self.cache_file.exists():
-            self.cache_file.unlink()
         if self.cache_file.exists():
+            if self.recompute:
+                self.cache_file.unlink()
             try:  # load
                 with self.cache_file.open('rb') as file:
-                    return pickle.load(file)
-            except (PermissionError, TypeError, EOFError, AttributeError, pickle.UnpicklingError):
+                    info_cache, check = pickle.load(file)
+            except (PermissionError, TypeError, EOFError, AttributeError, pickle.UnpicklingError,
+                    ValueError):
                 self.cache_file.unlink()
                 raise
-        else:
-            info_cache = super()._compute()
-            try:  # store
-                with self.cache_file.open('wb') as file:
-                    pickle.dump(info_cache, file)
-            except (PermissionError, TypeError):
+            if objects_equal(check, self._compute_check_data()):
+                return info_cache
+            else:
                 self.cache_file.unlink()
-                raise
-            return info_cache
+        info_cache = super()._compute()
+        check = self._compute_check_data()
+        try:  # store
+            with self.cache_file.open('wb') as file:
+                pickle.dump((info_cache, check), file)
+        except (PermissionError, TypeError):
+            self.cache_file.unlink()
+            raise
+        return info_cache
 
 
 class SubDataset(Dataset):
